Will not work for others
I have a relative file. This image will not work for others copying this and the html file will have a bug when loading the img src. 
This is what the image looks like https://imgur.com/a/kFWT3. I could change the code to make an external path and to accept that URL but I have to go now